implemented delete and edit function implemented alerts using swal2 fixed population on some controllers fixed links on some html

+++ frontend/public/scripts/bedTypeScript.js
@@ -34,6 +34,7 @@ bedTypeDropDown.addEventListener('change', (event) => {
   } else {
     bedTypeContainerDisplay('block', 'flex')
     bedTypeContainerChange('EDIT', 'ADD', 'block')
+    setSelectedBedTypeData()
   }
 })
 
@@ -48,26 +49,61 @@ function bedTypeContainerChange(newVal, oldVal, display) {
                             Bed Type Price: <input type='Number' name='bedTypePrice'>`
   bedTypeContainer.innerHTML = containerContent
   bedTypeFormButtons.forEach((button) => {
-    if (button.innerHTML === `${oldVal}`) {
+    if (button.id === `${oldVal}`) {
+      button.id = `${newVal}`
       button.innerHTML = `${newVal}`
-    } else if (button.innerHTML === 'DELETE') {
+    } else if (button.id === 'DELETE') {
       button.setAttribute('style', `display:${display}`)
+    } else {
     }
   })
-  const deleteButton = document.querySelector(
-    '.newroom-btn-group.bed-type > a:nth-child(3) > button'
+  const deleteButton = document.querySelector('.bed-type #DELETE')
+  const addButton = document.querySelector('.bed-type #ADD')
+  const editButton = document.querySelector('.bed-type #EDIT')
+
+  if (deleteButton) {
+    deleteButton.addEventListener('click', (e) => {
+      e.preventDefault()
+      deleteBedType()
+    })
+  }
+  if (addButton) {
+    addButton.addEventListener('click', (e) => {
+      e.preventDefault()
+      createBedType()
+    })
+  }
+  if (editButton) {
+    editButton.addEventListener('click', (e) => {
+      e.preventDefault()
+      editBedType()
+    })
+  }
+}
+
+async function setSelectedBedTypeData() {
+  const bedTypeNameInputElement = document.querySelector(
+    "input[name='bedTypeName']"
   )
-  const addButton = document.querySelector(
-    '.newroom-btn-group.bed-type > a:nth-child(4) > button'
+  const bedTypeSizeInputElement = document.querySelector(
+    "input[name='bedTypeSize']"
   )
-  deleteButton.addEventListener('click', (e) => {
-    e.preventDefault()
-    deleteBedType()
-  })
-  addButton.addEventListener('click', (e) => {
-    e.preventDefault()
-    createBedType()
-    form.reset()
+  const bedTypePriceInputElement = document.querySelector(
+    "input[name='bedTypePrice']"
+  )
+  const selectedBedIdName = bedTypeDropDown.value
+
+  const selectedBedInfo = await getData('http://localhost:3000/api/beds/types')
+
+  selectedBedInfo.forEach((data) => {
+    if (data.bedTypeName == selectedBedIdName) {
+      bedTypeNameInputElement.value =
+        data.bedTypeName ?? 'Bed type name does not exist'
+      bedTypeSizeInputElement.value =
+        data.bedTypeSize ?? 'Bed type size does not exist'
+      bedTypePriceInputElement.value =
+        data.bedTypePrice ?? 'Bed type price does not exist'
+    }
   })
 }
 
@@ -82,27 +118,39 @@ async function createBedType() {
     "input[name='bedTypePrice']"
   ).value
 
-  try {
-    const dataPost = {
-      bedTypeName: bedTypeNameValue,
-      bedTypeSize: bedTypeSizeValue,
-      bedTypePrice: bedTypePriceValue,
-    }
-
-    const response = await postData(
-      'http://localhost:3000/api/beds/types',
-      dataPost
-    )
-
-    if (response.message) {
-      return swal('Error!', `${response.message}`, 'error')
-    }
-
-    swal('Success', 'Room successfully added', 'success')
-  } catch (error) {
-    swal('Error!', `${error}`, 'error')
+  const dataPost = {
+    bedTypeName: bedTypeNameValue,
+    bedTypeSize: bedTypeSizeValue,
+    bedTypePrice: bedTypePriceValue,
   }
+
+  const response = await postData(
+    'http://localhost:3000/api/beds/types',
+    dataPost
+  )
+
+  if (response.message) {
+    return Swal.fire({
+      icon: 'error',
+      title: 'Error',
+      text: `${response.message}`,
+      showConfirmButton: true,
+      confirmButtonColor: '#ff2e63',
+    })
+  }
+
+  Swal.fire({
+    icon: 'success',
+    iconColor: '#54bab9',
+    title: 'Success!',
+    text: 'Bed type successfully Added',
+    showConfirmButton: true,
+    confirmButtonColor: '#ff2e63',
+  }).then((result) => {
+    location.reload()
+  })
 }
+
 async function deleteBedType() {
   const bedTypeResponse = await getData('http://localhost:3000/api/beds/types')
   let bedTypeObjectId
@@ -118,14 +166,77 @@ async function deleteBedType() {
     `${bedTypeObjectId}`
   )
   if (deleteResponse.message) {
-    return swal('Error!', `${deleteResponse.message}`, 'error')
+    return Swal.fire({
+      icon: 'error',
+      title: 'Error',
+      text: `${error.message}`,
+      showConfirmButton: true,
+      confirmButtonColor: '#ff2e63',
+    })
   }
 
-  swal('Success', 'Bed type successfully deleted!', 'success')
-
-  refreshPage()
+  Swal.fire({
+    icon: 'success',
+    iconColor: '#54bab9',
+    title: 'Success!',
+    text: 'Bed type successfully deleted',
+    showConfirmButton: true,
+    confirmButtonColor: '#ff2e63',
+  }).then((result) => {
+    location.reload()
+  })
 }
 
-function refreshPage() {
-  location.href = 'http://localhost:3000/rooms(bedTypeManage).html'
+async function editBedType() {
+  const bedTypeResponse = await getData('http://localhost:3000/api/beds/types')
+  let bedTypeObjectId
+  bedTypeResponse.forEach((data) => {
+    if (data.bedTypeName) {
+      if (data.bedTypeName == bedTypeDropDown.value) {
+        bedTypeObjectId = data._id
+      }
+    }
+  })
+
+  const bedTypeNameValue = document.querySelector(
+    "input[name='bedTypeName']"
+  ).value
+  const bedTypeSizeValue = document.querySelector(
+    "input[name='bedTypeSize']"
+  ).value
+  const bedTypePriceValue = document.querySelector(
+    "input[name='bedTypePrice']"
+  ).value
+
+  const dataPost = {
+    bedTypeName: bedTypeNameValue,
+    bedTypeSize: bedTypeSizeValue,
+    bedTypePrice: bedTypePriceValue,
+  }
+
+  const updateResponse = await putData(
+    'http://localhost:3000/api/beds/types/',
+    `${bedTypeObjectId}`,
+    dataPost
+  )
+  if (updateResponse.message) {
+    return Swal.fire({
+      icon: 'error',
+      title: 'Error',
+      text: `${error.message}`,
+      showConfirmButton: true,
+      confirmButtonColor: '#ff2e63',
+    })
+  }
+
+  Swal.fire({
+    icon: 'success',
+    iconColor: '#54bab9',
+    title: 'Success!',
+    text: 'Bed type successfully updated!',
+    showConfirmButton: true,
+    confirmButtonColor: '#ff2e63',
+  }).then((result) => {
+    location.reload()
+  })
 }
